src/requirements.txt
adafruit-circuitpython-ina219
adafruit-blinka
board
busio
RPI.GPIO
smbus2
